refactor: 코드 컨벤션 보완 및 파일 구조도에 맞춰 이동
