+++ package.json
@@ -27,5 +27,5 @@
     "tailwindcss": "^4",
     "typescript": "^5"
   },
-  "packageManager": "yarn@4.2.2"
+  "packageManager": "yarn@4.9.2"
 }
