+++ src/types/next-auth.d.ts
@@ -1,0 +1,11 @@
+import { DefaultSession } from 'next-auth';
+
+declare module 'next-auth' {
+  interface User {
+    id: string;
+  }
+
+  interface Session extends DefaultSession {
+    user?: User;
+  }
+}
